feat(style): upgrade ionicons
This removes the dated version of Ionicons and upgrades to latest, which 
now injects itself using a script (not sure how we feel about this - as 
when I disabled JS, the icon simply no longer render).

Anyway, I know there was discussion about moving off of Ionicons and 
leveraging the same icon set in our product suite, but this was easy and 
could tide us over for a while.