fix(links): make all links relative
I'm working on a simple buildkite pipeline that will build a blog and 
"deploy" the static files to s3 and allow for validation (and eventually 
integ testing?) - but I was running into issues given that I'm relying 
on a sub-path for my s3 path.

So, for my temp build, I'm planning on making a dynamic config file that 
will have something like this:

```yml
baseurl: "/developer-blog/18/_site"
```

But when I did that, I realized many of the exisiting links and elements 
in the theme were hard-coded to assume the blog was exposed at a root 
path.